Feat/add move override (#31)
* almost done

* wip refactor

* wip

* add move override done

+++ src/cpu/PlayerCPU.sol
@@ -15,7 +15,7 @@ contract PlayerCPU is CPU {
 
     constructor(uint256 numMoves, IEngine engine, ICPURNG rng) CPU(numMoves, engine, rng) {}
 
-    function setMove(bytes32 battleKey, uint256 moveIndex, bytes calldata extraData) external {
+    function setMove(bytes32 battleKey, uint128 moveIndex, bytes calldata extraData) external {
         if (msg.sender != ENGINE.getPlayersForBattle(battleKey)[0]) {
             revert NotP0();
         }
@@ -30,7 +30,7 @@ contract PlayerCPU is CPU {
         external
         view
         override
-        returns (uint256 moveIndex, bytes memory extraData)
+        returns (uint128 moveIndex, bytes memory extraData)
     {
         return (declaredMoveForBattle[battleKey].moveIndex, declaredMoveForBattle[battleKey].extraData);
     }
